update avatar update and coverImage update and Apply delete comments and likes on deletion of video

+++ src/controllers/video.controller.js
@@ -1,6 +1,8 @@
 import mongoose, { isValidObjectId } from "mongoose";
 import { Video } from "../models/video.model.js";
 import { User } from "../models/user.model.js";
+import { Comment } from "../models/comment.model.js"
+import { Like } from "../models/like.model.js";
 import { ApiError } from "../utils/ApiError.js";
 import { ApiResponse } from "../utils/ApiResponse.js";
 import { asyncHandler } from "../utils/asyncHandler.js";
@@ -250,6 +252,14 @@ const deleteVideo = asyncHandler(async (req, res) => {
   if (!thumbnailDeleteResponse || !videoDeleteResponse)
     throw new ApiError(500, "Failed to delete video");
 
+  const commentsDeleteResponse = await Comment.deleteMany({ video: videoId });
+  if(!commentsDeleteResponse)
+    throw new ApiError(500, "Failed to delete videos comments");
+
+  const likesDeleteResponse = await Like.deleteMany({ video: videoId });
+  if(!likesDeleteResponse)
+    throw new ApiError(500, "Failed to delete videos likes");
+  
   await Video.findByIdAndDelete(videoId);
 
   return res
